Convert olark flux store to a redux store
