--- client/lib/olark-store/constants.js
@@ -1,17 +1,0 @@
-/**
- * External dependencies
- */
-import keyMirror from 'key-mirror';
-
-/**
- * Module variables
- */
-export const action = keyMirror( {
-	OLARK_READY: null,
-	OLARK_OPERATORS_AWAY: null,
-	OLARK_OPERATORS_AVAILABLE: null,
-	OLARK_DETAILS: null,
-	OLARK_LOCALE: null,
-	OLARK_USER_ELIGIBILITY: null,
-	OLARK_SET_EXPANDED: null
-} );
