--- client/lib/olark-store/test/olark-store-test.js
@@ -1,19 +1,0 @@
-/**
- * External Dependencies
- */
-import { assert } from 'chai';
-import olarkStore from 'lib/olark-store';
-
-describe( 'olark-store', function() {
-	it( 'Olark store data should be an object', function() {
-		const data = olarkStore.get();
-		assert.isObject( data );
-	} );
-
-	it( 'Olark store data should have expected properties', function() {
-		const data = olarkStore.get();
-		assert.isBoolean( data.isOlarkReady );
-		assert.isBoolean( data.isOperatorAvailable );
-		assert.isObject( data.details );
-	} );
-} );
