+++ client/state/action-types.js
@@ -15,6 +15,13 @@ export const FETCH_SITE_PLANS = 'FETCH_SITE_PLANS';
 export const FETCH_SITE_PLANS_COMPLETED = 'FETCH_SITE_PLANS_COMPLETED';
 export const FETCH_WPORG_PLUGIN_DATA = 'FETCH_WPORG_PLUGIN_DATA';
 export const NEW_NOTICE = 'NEW_NOTICE';
+export const OLARK_DETAILS = 'OLARK_DETAILS';
+export const OLARK_LOCALE = 'OLARK_LOCALE';
+export const OLARK_OPERATORS_AVAILABLE = 'OLARK_OPERATORS_AVAILABLE';
+export const OLARK_OPERATORS_AWAY = 'OLARK_OPERATORS_AWAY';
+export const OLARK_READY = 'OLARK_READY';
+export const OLARK_SET_EXPANDED = 'OLARK_SET_EXPANDED';
+export const OLARK_USER_ELIGIBILITY = 'OLARK_USER_ELIGIBILITY';
 export const POSTS_RECEIVE = 'POSTS_RECEIVE';
 export const POSTS_REQUEST = 'POSTS_REQUEST';
 export const POSTS_REQUEST_FAILURE = 'POSTS_REQUEST_FAILURE';
