+++ client/state/olark/selector.js
@@ -1,0 +1,3 @@
+export function getOlarkState( state ) {
+	return state.olark.toJS();
+}
