+++ client/state/olark/test/lib/mock-dispatch.js
@@ -1,0 +1,3 @@
+export function dispatch( state ) {
+	return state;
+}
